+++ python/lsst/ts/m2com/mock/__init__.py
@@ -27,5 +27,6 @@ from .mock_control_open_loop import *
 from .mock_control_closed_loop import *
 from .mock_error_handler import *
 from .mock_power_system import *
+from .mock_inner_loop_controller import *
 from .mock_model import *
 from .mock_server import *